Add cancel command

+++ main.py
@@ -1,4 +1,6 @@
 import os
+import sys
+import traceback
 import time
 import shutil
 from pyrogram import Client, filters, types, enums
@@ -11,9 +13,7 @@ from keep_alive import keep_alive
 keep_alive()
 
 
-apiid = os.getenv("API_ID")
-bottoken = os.getenv("BOT_TOKEN")
-apihash = os.getenv("API_HASH")
+me = os.getenv("ME")
 token = os.getenv("SESSION")
 # sesion de mi bot
 bot = Client("pyro_bot",session_string=token)
@@ -56,43 +56,93 @@ message_reply = "" # referencia de topic o grupo para enviar los mensajes de res
 message_chat = "" # referencia de chat donde se hizo el comando
 client_reference = "" # referencia del cliente para usar en diferentes funciones
 changeNameDoc = "" # referencia para cambiar nombres de documentos aportados
+cancel_words = ["cancel", "/cancel", "cancelar", "/cancelar"] # referencias de palabras para cancelar proceso de aporte en partes donde se deba response sino se debe usar el comando /cancel
+cancel_commands = ["cancel", "cancelar"] # referencias de comandos para cancelar aporte 
 
 # -------REVISAR----------
 #
 # -------REVISAR----------
 
+# -------CANCELAR----------
+@bot.on_message(filters.command("restart") & filters.user(me) & filters.private)
+async def asd(client, m):
+    try:
+        print("El bot se reiniciara espere...")
+        restart()
+    except Exception as e:
+        traceback.print_exc()
+
+@bot.on_message(filters.command("id"))
+async def asd(client, m):
+    try:
+        print(m.from_user.id)
+    except Exception as e:
+        traceback.print_exc()
+
+@bot.on_message(filters.command(cancel_commands))
+async def asd(client, m):
+    try:
+        global hand
+        chat_id = m.chat.id
+        print("se cancelara el aporte")
+        await bot.delete_messages(chat_id, m.id)
+        ne = await bot.get_messages(m.chat.id,m.id-1)
+        if ne.reply_markup:
+            print("se cancelo mensaje de repuestas para evitar errores")
+            bot.edit_message_reply_markup(chat_id, ne.id)
+        if hand:
+            print("se borro handler y se termino proceso de aporte")
+            await bot.delete_messages(chat_id, message_reference)
+            bot.remove_handler(*hand)
+            hand = ""
+    except Exception as e:
+        traceback.print_exc()
+
+
+def restart():
+    return os.execl(sys.executable, os.path.abspath(__file__), * sys.argv)
+# -------CANCELAR----------
+
 # aqui inicia el comando
 @bot.on_message(filters.command("aporte"))
 async def aportar(client,m):
-    global message_reference
-    global message_chat
-    global message_reply
-    global client_reference
+    try:
+        global message_reference
+        global message_chat
+        global message_reply
+        global client_reference
 
-    message_chat = m.chat
-    message_reply = m.reply_to_message_id if m.reply_to_message_id else 1
+        message_chat = m.chat
+        message_reply = m.reply_to_message_id if m.reply_to_message_id else 1
 
-    if not config.isDataOfGroup(str(message_chat.id)[4:]):
-        send = "Debes configurar tu lista de hastgas y tu canal de aportes primero, usa el comando /help para mas información."
-        await bot.send_message(message_chat.id, send, reply_to_message_id=message_reply)
-    else:
-        message_reference = m.id
-        client_reference = client
-        await my_handler()
+        if not config.isDataOfGroup(str(message_chat.id)[4:]):
+            send = "Debes configurar tu lista de hastgas y tu canal de aportes primero, usa el comando /help para mas información."
+            res = await bot.send_message(message_chat.id, send, reply_to_message_id=message_reply)
+            time.sleep(3)
+            await m.delete(), await res.delete()
+        else:
+            message_reference = m.id
+            client_reference = client
+            await my_handler()
+    except Exception:
+        traceback.print_exc()
 
 
 # funcion que crea botones de si y no
 def botonera(name, change=""):
-    if not change:
-        bto_pdf = types.InlineKeyboardButton("Si",callback_data=f'si_{name}')
-        bto_image = types.InlineKeyboardButton("No",callback_data=f'no_{name}')
-        botonera = types.InlineKeyboardMarkup([[bto_pdf,bto_image]])
-        return botonera
-    else:
-        bto_pdf = types.InlineKeyboardButton("Si",callback_data=f'{name}Si_{change}')
-        bto_image = types.InlineKeyboardButton("No",callback_data=f'{name}No_{change}')
-        botonera = types.InlineKeyboardMarkup([[bto_pdf,bto_image]])
-        return botonera
+    try:
+        if not change:
+            bto_pdf = types.InlineKeyboardButton("Si",callback_data=f'si_{name}')
+            bto_image = types.InlineKeyboardButton("No",callback_data=f'no_{name}')
+            botonera = types.InlineKeyboardMarkup([[bto_pdf,bto_image]])
+            return botonera
+        else:
+            bto_pdf = types.InlineKeyboardButton("Si",callback_data=f'{name}Si_{change}')
+            bto_image = types.InlineKeyboardButton("No",callback_data=f'{name}No_{change}')
+            botonera = types.InlineKeyboardMarkup([[bto_pdf,bto_image]])
+            return botonera
+    except Exception:
+        traceback.print_exc()
 
 
 # funcion para escuchar respuesta de si y no 
@@ -118,6 +168,10 @@ async def checker(client, call):
             respuestas[f"{res}_response"] = "si"
             send = "Escribe el nuevo nombre para tu archivo:"
             respu =  await bot.ask(message_chat.id, text=send, reply_to_message_id=message_reply)
+            if respu.text in cancel_words:
+                # return restart()
+                print("Se cancelo el aporte")
+                return
             nameDoc = respu.text
             respuestas[res] = nameDoc
             await respu.delete(), await respu.sent_message.delete()
@@ -141,9 +195,10 @@ async def checker(client, call):
                 await changeName()
             else:
                 await moreDocs()
-    except Exception as e:
-        print("error",e)
+    except Exception:
+        print("on_callback_query Error")
         await bot.delete_messages(call.message.chat.id,call.message.id)
+        traceback.print_exc()
 
 
 # funcion que guarda las respuestas de botones y ejecuta siguiente paso
@@ -181,50 +236,61 @@ async def my_handler(dat="",  response=""):
             await bot.delete_messages(message_chat.id,message_reference)
             message_reference = res.id
 
-    except Exception as e:
-        print("handler")
-        print(e)
+    except Exception:
+        print("handler Error")
+        traceback.print_exc()
 
 
 # funcion que ejecuta el siguiente paso luego de respuesta de boton
 async def actionBot(dato, metodo, action,  next=""):
-    global message_reference
-    global hand
-    if metodo == "send_message":
-        if type(action) == list: # mensajes de preguntas normales
-            mes = await bot.send_message(message_chat.id, action[0],reply_markup=botonera(action[1]), reply_to_message_id=message_reply)
-            message_reference = mes.id
-        else: # mensaje final,aqui se inicia la funcion para enviar el aporte al canal ya establecido anteriormente.
-            mes = await bot.send_message(message_chat.id, action, reply_to_message_id=message_reply)
-            message_reference = mes.id
-            time.sleep(2)
-            await bot.delete_messages(message_chat.id, message_reference)
-            esperando = await bot.send_message(message_chat.id,f"Ahora enviare tu aporte, espera un momento...", reply_to_message_id=message_reply)
-            await callE(str(message_chat.id)[4:])
-            await bot.delete_messages(message_chat.id,esperando.id)
+    try:
+        global message_reference
+        global hand
+        if metodo == "send_message":
+            if type(action) == list: # mensajes de preguntas normales
+                mes = await bot.send_message(message_chat.id, action[0],reply_markup=botonera(action[1]), reply_to_message_id=message_reply)
+                message_reference = mes.id
+            else: # mensaje final,aqui se inicia la funcion para enviar el aporte al canal ya establecido anteriormente.
+                mes = await bot.send_message(message_chat.id, action, reply_to_message_id=message_reply)
+                message_reference = mes.id
+                time.sleep(2)
+                await bot.delete_messages(message_chat.id, message_reference)
+                esperando = await bot.send_message(message_chat.id,f"Ahora enviare tu aporte, espera un momento...", reply_to_message_id=message_reply)
+                await isConfig(str(message_chat.id)[4:])
+                await bot.delete_messages(message_chat.id,esperando.id)
 
-    if metodo == "ask":
-        res = await bot.send_message(message_chat.id, action, reply_to_message_id=message_reply)
-        has = await res.chat.listen()
-        await bot.delete_messages(message_chat.id,message_reference)
-        respuestas[dato] = has.text
+        if metodo == "ask":
+            # res = await bot.send_message(message_chat.id, action, reply_to_message_id=message_reply)
+            # has = await res.chat.listen()
+            # --------------
+            res = await bot.ask(message_chat.id, action, reply_to_message_id=message_reply)
+            print(res.text)
+            if res.text in cancel_words:
+                # return restart()
+                print("se cancelo el aporte en ask")
+                return
+            has = res
+            await bot.delete_messages(message_chat.id,message_reference)
+            respuestas[dato] = has.text
 
-        await bot.delete_messages(message_chat.id,[res.id,has.id])
-        if next:
-            await my_handler(next)
+            await bot.delete_messages(message_chat.id,[res.id,has.id])
+            if next:
+                await my_handler(next)
 
-    if metodo == "next":
-        await bot.delete_messages(message_chat.id,message_reference)
-        res = await bot.send_message(message_chat.id, action[0], reply_to_message_id=message_reply)
-        message_reference = res.id
-        be = bot.add_handler(MessageHandler(func[action[1]][0], func[action[1]][1]))
-        hand = be
+        if metodo == "next":
+            await bot.delete_messages(message_chat.id,message_reference)
+            res = await bot.send_message(message_chat.id, action[0], reply_to_message_id=message_reply)
+            message_reference = res.id
+            be = bot.add_handler(MessageHandler(func[action[1]][0], func[action[1]][1]))
+            hand = be
+    except Exception:
+        traceback.print_exc()
 
 
 # verificador de documentos 
 async def verify_doc(client,message):
-    global listDoc
     try:
+        global listDoc
         mes_id = message.id
         # if message.document:# aqui se guardara el documento para enviarse al final de flujo
         listDoc.append(message)
@@ -236,26 +302,31 @@ async def verify_doc(client,message):
             bot.remove_handler(*hand)
             hand = ""
             await edit_docs()
-    except Exception as e:
-        print("verif error", e)
+    except Exception:
+        print("verify_doc Error")
+        traceback.print_exc()
 
 
 # añadiendo handler para esperar la imagen de referencia.
 async def verify_img(client,message):
-    global hand
-    # print("verify_img message.id: ",message.id)
-    if message.photo: # guarda imagen de referencia para enviar despues
-        photoId = message.photo.file_id
-        # aqui guardamos la imagen
-        neo = await bot.download_media(photoId, in_memory=True)
-        respuestas["imageDoc"] = neo
-        # eliminaresmos la el mensaje con la foto y el mensaje anterior para que este mas limpio
-        await bot.delete_messages(message_chat.id, [message_reference, message.id] )
-        # se debe remover el escuchador de imagenes
-        bot.remove_handler(*hand)
-        hand = ""
-        # despues de esto seguira el flujo de documento
-        await my_handler("document")
+    try:
+        global hand
+        # print("verify_img message.id: ",message.id)
+        if message.photo: # guarda imagen de referencia para enviar despues
+            photoId = message.photo.file_id
+            # aqui guardamos la imagen
+            neo = await bot.download_media(photoId, in_memory=True)
+            respuestas["imageDoc"] = neo
+            # eliminaresmos la el mensaje con la foto y el mensaje anterior para que este mas limpio
+            await bot.delete_messages(message_chat.id, [message_reference, message.id] )
+            # se debe remover el escuchador de imagenes
+            bot.remove_handler(*hand)
+            hand = ""
+            # despues de esto seguira el flujo de documento
+            await my_handler("document")
+    except Exception:
+        print("verify_img Error")
+        traceback.print_exc()
 
 
 func = {
@@ -266,8 +337,8 @@ func = {
 
 # revisara si hay mas mensajes que se enviaron
 async def revisar(mes):
-    ne = await bot.get_messages(message_chat.id,mes+1)
     try:
+        ne = await bot.get_messages(message_chat.id,mes+1)
         if ne.empty:
             global hand
             bot.remove_handler(*hand)
@@ -275,8 +346,9 @@ async def revisar(mes):
             # print(f"revisar - Se iniciara el editor de archivos")
             await bot.delete_messages(message_chat.id, [mes] )
             await edit_docs()
-    except Exception as e:
-        print("Ocurrio algo en revisar", e)
+    except Exception:
+        print("Revisar Error")
+        traceback.print_exc()
         send = "Ocurrio un error por favor informa al equipo o intentalo más tarde."
         res = await bot.send_message(message_chat.id, send, reply_to_message_id=message_reply)
         time.sleep(5)
@@ -308,40 +380,52 @@ async def edit_docs():
         await changeName()
 
         # listDoc = []
-    except Exception as error:
-        print("edit docs error")
-        print(error)
+    except Exception:
+        print("edit docs Error")
+        traceback.print_exc()
 
 
 async def moreDocs():
-    global respuestas
-    message = "¿ Quieres enviar mas documentos ?"
-    await bot.send_message(message_chat.id, message, reply_markup=botonera("moreDocs"), reply_to_message_id=message_reply)
-    await bot.delete_messages(message_chat.id, message_reference)
+    try:
+        global respuestas
+        message = "¿ Quieres enviar mas documentos ?"
+        await bot.send_message(message_chat.id, message, reply_markup=botonera("moreDocs"), reply_to_message_id=message_reply)
+        await bot.delete_messages(message_chat.id, message_reference)
+    except Exception:
+        print("moreDocs Error")
+        traceback.print_exc()
 
 
 async def changeName():
-    await bot.delete_messages(message_chat.id, message_reference)
-    send = f"¿ Quieres ponerle un nuevo nombre al archivo <b>{listDoc[0].document.file_name}</b> ?"
-    await bot.send_message(message_chat.id, send, reply_markup=botonera("changeName",listDoc[0].document.file_unique_id), reply_to_message_id=message_reply)
+    try:
+        await bot.delete_messages(message_chat.id, message_reference)
+        send = f"¿ Quieres ponerle un nuevo nombre al archivo <b>{listDoc[0].document.file_name}</b> ?"
+        await bot.send_message(message_chat.id, send, reply_markup=botonera("changeName",listDoc[0].document.file_unique_id), reply_to_message_id=message_reply)
+    except Exception:
+        print("changeName Error")
+        traceback.print_exc()
 
 
 # comando para enviar el aporte formateado.
-async def callE(chat_id):
-    global hand
-    if config.isDataOfGroup(chat_id):
-        # pasa hcaer la funcion de aporte
-        await sand()
-    else:
-        send = "Necesitas configurar tu lista de hastags y tu canal de aportes para poder mandar tu aporte, usa los comandos /start /createHash /setTopic" + "\n"
-        send += "Luego usa el comando /send para enviar el aporte que acabas de hacer"
-        await bot.send_message(message_chat.id, send)
-        hand = bot.add_handler(MessageHandler(sand(message_chat.id),filters.command("send")))
-        # manda mensaje de que necesitas configurar tu lista de hastags y tu canal de aportes
+async def isConfig(chat_id):
+    try:
+        global hand
+        if config.isDataOfGroup(chat_id):
+            # pasa hcaer la funcion de aporte
+            await send_aport()
+        else:
+            send = "Necesitas configurar tu lista de hastags y tu canal de aportes para poder mandar tu aporte, usa los comandos /start /createHash /setTopic" + "\n"
+            send += "Luego usa el comando /send para enviar el aporte que acabas de hacer"
+            await bot.send_message(message_chat.id, send)
+            hand = bot.add_handler(MessageHandler(send_aport(message_chat.id),filters.command("send")))
+            # manda mensaje de que necesitas configurar tu lista de hastags y tu canal de aportes
+    except Exception:
+        print("isConfig Error")
+        traceback.print_exc()
 
 
 # funcion que envia el aporte formateado
-async def sand():
+async def send_aport():
     try:
         hastags = [] 
         text_hastags = ""
@@ -396,9 +480,9 @@ async def sand():
         respuestas = {}
         # print("sand -Termine aporte")
 # ------------------------------------
-    except Exception as error:
-        print("error snad")
-        print(error)
+    except Exception:
+        print("isConfig send_aport")
+        traceback.print_exc()
 
 
 bot.run(print("me", "ya esya"))
